Manual control added, moved to separate file
- Added functions to add one minute or hour to the current time
- Fixed errors trying to do math with an unsigned long and an int
- Moved to Timekeeping.h to keep the main file clean.

+++ digitalAlarm/Constants.h
@@ -1,3 +1,3 @@
-constexpr int msPerHour{3600000};
-constexpr int msPerMinute{60000};
-constexpr int msPerSecond{1000};
+constexpr unsigned long msPerHour{3600000};
+constexpr unsigned long msPerMinute{60000};
+constexpr unsigned long msPerSecond{1000};
